+++ src/common/util.ts
@@ -1,9 +1,13 @@
 import * as vscode from 'vscode';
 
+enum Confirm {
+    YES = "YES", NO = "NO"
+}
+
 export class Util {
     public static buildPath(data: string, webview: vscode.Webview, contextPath: string): string {
         return data.replace(/((src|href)=("|')?)(\/\/)/gi, "$1http://")
-            .replace(/((src|href)=("|'))((?!(http))[^"']+?\.(css|js|properties|json|png|jpg))\b/gi, "$1" + webview.asWebviewUri(vscode.Uri.file(`${contextPath}`)) + "/$4");
+            .replace(/((src|href)=("|'))((?!(http|#)).+?["'])/gi, "$1" + webview.asWebviewUri(vscode.Uri.file(`${contextPath}`)) + "/$4");
     }
 
     public static listen(webviewPanel: vscode.WebviewPanel, uri: vscode.Uri, callback: () => void, disposeCallback?: () => void) {
@@ -13,4 +17,18 @@ export class Util {
         });
     }
 
+    public static async confirm(title: string, placeHolder: string, callback?: () => void): Promise<boolean> {
+        return this.confirmActual({ title, placeHolder }, callback)
+    }
+
+    public static async confirmActual(options: vscode.QuickPickOptions, callback?: () => void): Promise<boolean> {
+        const res = await vscode.window.showQuickPick([Confirm.YES, Confirm.NO], options);
+        const yes = res == Confirm.YES;
+        if (yes && callback) {
+            await callback()
+            return true;
+        }
+        return yes;
+    }
+  
 }