+++ src/service/markdown/markdown-pdf.js
@@ -1,13 +1,8 @@
 const fs = require("fs")
-const os = require("os")
-const puppeteer = require("puppeteer-core")
 const path = require("path")
-const cheerio = require("cheerio")
 const url = require("url")
 const URI = require("vscode").Uri
-const hljs = require("highlight.js")
 const markdownIt = require("markdown-it")
-const mustache = require("mustache")
 const markdownItCheckbox = require("markdown-it-checkbox")
 const markdownItKatex = require("./ext/markdown-it-katex")
 const markdownItPlantuml = require("markdown-it-plantuml")
@@ -38,16 +33,17 @@ function addTocToContent(text) {
  * convert markdown to html (markdown-it)
  */
 function convertMarkdownToHtml(filename, type, text, config) {
-  text = addTocToContent(text)
+  if (type == 'pdf') text = addTocToContent(text)
   let md = {}
 
   try {
     try {
+      const hljs = require("highlight.js");
       console.log("[pretty-md-pdf] Converting (convertMarkdownToHtml) ...")
-      let breaks = config["breaks"]
+      const breaks = config["breaks"]
       md = markdownIt({
         html: true,
-        breaks: breaks,
+        breaks,
         highlight: function (str, lang) {
           if (lang && hljs.getLanguage(lang)) {
             try {
@@ -88,7 +84,7 @@ function convertMarkdownToHtml(filename, type, text, config) {
       // convert the img src of the html
       md.renderer.rules.html_block = function (tokens, idx) {
         let html = tokens[idx].content
-        let $ = cheerio.load(html)
+        let $ = require("cheerio").load(html)
         $("img").each(function () {
           let src = $(this).attr("src")
           let href = convertImgPath(src, filename)
@@ -117,6 +113,7 @@ function convertMarkdownToHtml(filename, type, text, config) {
  */
 function mergeHtml(content, uri) {
   try {
+    const mustache = require("mustache")
     const title = path.basename(uri.fsPath)
     const style = readStyles()
     const templatePath = path.join(__dirname, "template", "template.html")
@@ -139,40 +136,7 @@ function isExistsPath(path) {
   }
 }
 
-function isExistsDir(dirname) {
-  if (dirname.length === 0) {
-    return false
-  }
-  try {
-    if (fs.statSync(dirname).isDirectory()) {
-      return true
-    } else {
-      console.warn("Directory does not exist!")
-      return false
-    }
-  } catch (error) {
-    console.warn(error.message)
-    return false
-  }
-}
-
-function getFolder(resource) {
-  return {
-    index: 0,
-    name: path.basename(resource.path),
-    uri: URI.file(path.dirname(resource.path))
-  }
-}
-
-
-
-function readFile(filename, encode) {
-  if (filename.length === 0) {
-    return ""
-  }
-  if (!encode && encode !== null) {
-    encode = "utf-8"
-  }
+function readFile(filename, encode = "utf-8") {
   if (filename.indexOf("file://") === 0) {
     if (process.platform === "win32") {
       filename = filename.replace(/^file:\/\/\//, "")
@@ -240,72 +204,6 @@ function readStyles() {
   }
 }
 
-function checkPuppeteerBinary(config) {
-  try {
-    // settings.json
-    let executablePath = config["executablePath"] || ""
-    if (isExistsPath(executablePath)) {
-      return true
-    }
-
-    // bundled Chromium
-    executablePath = puppeteer.executablePath()
-    if (isExistsPath(executablePath)) {
-      return true
-    } else {
-      return false
-    }
-  } catch (error) {
-    showErrorMessage("checkPuppeteerBinary()", error)
-  }
-}
-
-/*
- * puppeteer install.js
- * https://github.com/GoogleChrome/puppeteer/blob/master/install.js
- */
-async function installChromium(config) {
-  try {
-    console.log("[pretty-md-pdf] Installing Chromium ...")
-
-    // proxy setting
-    setProxy(config)
-
-    const puppeteerMetadata = require(path.join(__dirname, "node_modules", "puppeteer", "package.json"))
-
-    let revision = puppeteerMetadata.puppeteer.chromium_revision
-    let browserFetcher = puppeteer.createBrowserFetcher()
-    let revisionInfo = browserFetcher.revisionInfo(revision)
-
-    // download Chromium
-    try {
-      await browserFetcher.download(revisionInfo.revision, (downloadedBytes, totalBytes) => {
-        let progress = parseInt(downloadedBytes / totalBytes * 100)
-        console.log("[pretty-md-pdf] Installing Chromium " + progress + "%")
-      })
-    } catch (ex) {
-      console.log("[pretty-md-pdf] ERROR: Failed to download Chromium!")
-    }
-
-    console.log("Chromium downloaded to " + revisionInfo.folderPath)
-
-    let localRevisions = await browserFetcher.localRevisions()
-    localRevisions = localRevisions.filter(revision => revision !== revisionInfo.revision)
-    // Remove previous chromium revisions.
-    let cleanupOldVersions = localRevisions.map(revision => browserFetcher.remove(revision))
-
-    if (checkPuppeteerBinary(config)) {
-
-      console.log("[pretty-md-pdf] Chromium installation succeeded!")
-      console.log("[pretty-md-pdf] Chromium installation succeeded.")
-
-      await Promise.all(cleanupOldVersions)
-    }
-  } catch (error) {
-    showErrorMessage("installChromium()", error)
-  }
-}
-
 function showErrorMessage(msg, error) {
   console.error("ERROR: " + msg)
   console.log("ERROR: " + msg)
@@ -315,30 +213,10 @@ function showErrorMessage(msg, error) {
   }
 }
 
-function setProxy(config) {
-  let https_proxy = config["proxy"] || ""
-  if (https_proxy) {
-    process.env.HTTPS_PROXY = https_proxy
-    process.env.HTTP_PROXY = https_proxy
-  }
-}
-
-async function init(config) {
-  try {
-    if (checkPuppeteerBinary(config)) {
-    } else {
-      await installChromium(config)
-    }
-  } catch (error) {
-    showErrorMessage("init()", error)
-  }
-}
-
 export const convertMd = async (options) => {
   const config = options.config
   options.outputFileType = config.type[0]
   console.log(`[pretty-md-pdf] Converting markdown file: ${options.markdownFilePath}`)
-  await init(config)
   await convertMarkdown(
     path.resolve(options.markdownFilePath),
     config